Allow watchdog to function correctly.
The watchdog was resetting too often, not allowing the firmware to run for extended periods of time. Under normal circumstances, the watchdog is reset every second, with a reset period of roughly 2 seconds. This is fine, as the watchdog resets before its period is hit. However, when UPSat starts to transmit CW data, the transmit process takes roughly 10 seconds, delaying the refresh and causing the watchdog to expire.

The solution was to extend the reset period for the watchdog. Previously, the reset value used a unit32_t type, which only allowed a max value of 4 seconds. This seems almost contradictory with the IWDG implementation we ported over, since their period seems to exceed was can be stored in 32 bits (and there's a mismatch between some 32 and 64 bit data types). Changing the period to use a unit64_t type allows for the period to extend to roughly 11 seconds, giving enough time for the telemetry data to transmit.

This is hopefully enough for our purposes, because otherwise we'll have to figure out why the CW transmit takes longer than the watchdog reset period, preventing the refresh function to trigger in time.